payment seller main

+++ src/main/java/com/poly/DATN_BookWorms/entities/Bookings.java
@@ -65,7 +65,7 @@ public class Bookings {
 
 	public Double     costvoucher ;
 
-	public Double     costship ;
+	public Float     costship ;
 
     //--- ENTITY LINKS ( RELATIONSHIP )
     @OneToMany(mappedBy="bookings")
